+++ sydtest/test/Test/Syd/GoldenSpec.hs
@@ -14,6 +14,18 @@ spec = do
     it "outputs the same as last time" $ do
       pureGoldenTextFile
         "test_resources/output.golden"
-        (LT.toStrict $ LTB.toLazyText $ renderResultReport defaultSettings With24BitColours (Timed [] 0))
+        ( LT.toStrict $
+            LTB.toLazyText $
+              renderResultReport
+                defaultSettings
+                With24BitColours
+                ( Timed
+                    { timedValue = [],
+                      timedWorker = 0,
+                      timedBegin = 0,
+                      timedEnd = 0
+                    }
+                )
+        )
   describe "defaultSettings" $ do
     it "is the same thing as last time" $ goldenPrettyShowInstance "test_resources/defaultSettings-show.golden" defaultSettings